universal7870: audio: fix audioserver crash
The following happens as we had 2 libaudioroute.so:

- one in /system (build by system/media/audio_route)
- and the other in /vendor (copied as blob)

> E vndksupport: Could not load /vendor/lib/hw/audio.primary.universal7870.so from default namespace: dlopen failed: cannot locate symbol "audio_route_get_mixer_ctl" referenced by "/system/vendor/lib/hw/audio.primary.universal7870.so"....

Removing one of the both causes this error:

> E vndksupport: Could not load /vendor/lib/hw/audio.primary.universal7870.so from default namespace: dlopen failed: cannot locate symbol "mixer_update" referenced by "/system/lib/libaudioroute.so"....

I was not able to find a blob or ref to this symbol
(likely I'm just blind) so I decided to shim this.

Shimming seems to work fine while not very intensively
tested - but at least audio is back.

Signed-off-by: steadfasterX <steadfasterX@gmail.com>