universal7870: selinux: fix first call no audio
I was not able to set:

> allow system_suspend hal_audio_default:binder transfer;

as hal_audio_default is part of vendor which is not avail in private
but system_suspend is part of private and not avail in vendor.

this commit fixes the first call on boot issue until I find a better
way fixing that. without this the very first call will loose audio after
some seconds while subsequent calls work fine.

log:

> I auditd  : type=1400 audit(0.0:250): avc: denied { transfer } for comm="suspend@1.0-ser" scontext=u:r:system_suspend:s0 tcontext=u:r:hal_audio_default:s0 tclass=binder permissive=1
> I suspend@1.0-ser: type=1400 audit(0.0:250): avc: denied { transfer } for scontext=u:r:system_suspend:s0 tcontext=u:r:hal_audio_default:s0 tclass=binder permissive=1

Signed-off-by: steadfasterX <steadfasterX@gmail.com>